+++ modules/project/exec/src/sap/commerce/toolset/project/runConfigurations/LocalSapCXConfigurationType.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -22,10 +22,11 @@ import com.intellij.execution.configurations.RunConfiguration
 import com.intellij.execution.configurations.SimpleConfigurationType
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.NotNullLazyValue
+import sap.commerce.toolset.HybrisConstants
 import sap.commerce.toolset.HybrisIcons
 
 class LocalSapCXConfigurationType : SimpleConfigurationType(
-    "SAPCX",
+    HybrisConstants.RunConfiguration.SAP_CX,
     "SAP CX Start",
     "Start local SAP CX Platform Configuration Type",
     NotNullLazyValue.createValue { HybrisIcons.Y.LOGO_BLUE }