+++ modules/shared/core/src/sap/commerce/toolset/HybrisConstants.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -302,5 +302,11 @@ object HybrisConstants {
     object Folding {
         const val NO_VALUE = "🚫" // ? or 🚫 ∅
         const val DESCRIPTION_PREFIX = "ⓘ"
+    }
+
+    object RunConfiguration {
+        // see JUnitConfigurationType
+        const val JUNIT = "JUnit"
+        const val SAP_CX = "SAPCX"
     }
 }